Add game function and final message
OK so I got confused and completely forgot to commit anything. I was
heavily struggling with the looping aspect and how to add scores but by
the end of it I got the gist of it. I think. The code could probably be
cleaner in some aspects but I'm happy it worked out in the end.